+++ src/databricks/sql/utils.py
@@ -9,7 +9,6 @@ from decimal import Decimal
 from enum import Enum
 from typing import Any, Dict, List, Optional, Union
 import re
-from ssl import SSLContext, CERT_NONE, CERT_REQUIRED, create_default_context
 
 import lz4.frame
 import pyarrow
@@ -21,6 +20,7 @@ from databricks.sql.thrift_api.TCLIService.ttypes import (
     TSparkArrowResultLink,
     TSparkRowSetType,
 )
+from databricks.sql.types import SSLOptions
 
 from databricks.sql.parameters.native import ParameterStructure, TDbsqlParameter
 
@@ -29,53 +29,6 @@ import logging
 BIT_MASKS = [1, 2, 4, 8, 16, 32, 64, 128]
 
 logger = logging.getLogger(__name__)
-
-
-class SSLOptions:
-    tls_verify: bool
-    tls_verify_hostname: bool
-    tls_trusted_ca_file: Optional[str]
-    tls_client_cert_file: Optional[str]
-    tls_client_cert_key_file: Optional[str]
-    tls_client_cert_key_password: Optional[str]
-
-    def __init__(
-        self,
-        tls_verify: Optional[bool] = True,
-        tls_verify_hostname: Optional[bool] = True,
-        tls_trusted_ca_file: Optional[str] = None,
-        tls_client_cert_file: Optional[str] = None,
-        tls_client_cert_key_file: Optional[str] = None,
-        tls_client_cert_key_password: Optional[str] = None,
-    ):
-        self.tls_verify = tls_verify
-        self.tls_verify_hostname = tls_verify_hostname
-        self.tls_trusted_ca_file = tls_trusted_ca_file
-        self.tls_client_cert_file = tls_client_cert_file
-        self.tls_client_cert_key_file = tls_client_cert_key_file
-        self.tls_client_cert_key_password = tls_client_cert_key_password
-
-    def create_ssl_context(self) -> SSLContext:
-        ssl_context = create_default_context(cafile=self.tls_trusted_ca_file)
-
-        if self.tls_verify is False:
-            ssl_context.check_hostname = False
-            ssl_context.verify_mode = CERT_NONE
-        elif self.tls_verify_hostname is False:
-            ssl_context.check_hostname = False
-            ssl_context.verify_mode = CERT_REQUIRED
-        else:
-            ssl_context.check_hostname = True
-            ssl_context.verify_mode = CERT_REQUIRED
-
-        if self.tls_client_cert_file:
-            ssl_context.load_cert_chain(
-                certfile=self.tls_client_cert_file,
-                keyfile=self.tls_client_cert_key_file,
-                password=self.tls_client_cert_key_password,
-            )
-
-        return ssl_context
 
 
 class ResultSetQueue(ABC):
